test: export `.hoodieApi()` factory method
