+++ tests/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-require('./specs/constructor')
+require('./specs/factory')
 require('./specs/add')
 require('./specs/find')
 require('./specs/find-all')
