+++ tests/specs/find-or-add.js
@@ -3,13 +3,12 @@
 var test = require('tape')
 
 var dbFactory = require('../utils/db')
-var Store = require('../../')
 
 test('store.findOrAdd exists', function (t) {
   t.plan(1)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   t.is(typeof store.findOrAdd, 'function', 'has method')
 })
@@ -18,7 +17,7 @@ test('hoodie.store.findOrAdd(id, object) finds existing', function (t) {
   t.plan(2)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   store.add({id: 'exists', foo: 'bar'})
 
@@ -35,7 +34,7 @@ test('hoodie.store.findOrAdd(id, object) adds new', function (t) {
   t.plan(2)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   return store.findOrAdd('newid', {foo: 'baz'})
 
@@ -49,7 +48,7 @@ test('hoodie.store.findOrAdd(id) fails if no object exists', function (t) {
   t.plan(1)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   store.findOrAdd('thing')
   .catch(function (error) {
@@ -61,7 +60,7 @@ test('hoodie.store.findOrAdd(object) finds existing', function (t) {
   t.plan(2)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   store.add({id: 'exists', foo: 'bar'})
 
@@ -79,7 +78,7 @@ test('hoodie.store.findOrAdd(object) adds new', function (t) {
   t.plan(2)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   return store.findOrAdd({id: 'newid', foo: 'baz'})
 
@@ -93,7 +92,7 @@ test('hoodie.store.findOrAdd(object) fails if object has no id', function (t) {
   t.plan(1)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   return store.findOrAdd({foo: 'bar'})
 
@@ -106,7 +105,7 @@ test('hoodie.store.findOrAdd([object1, object2])', function (t) {
   t.plan(4)
 
   var db = dbFactory()
-  var store = new Store(db)
+  var store = db.hoodieApi()
 
   store.add([
     {id: 'exists'}
