+++ tests/utils/db.js
@@ -1,6 +1,7 @@
 'use strict'
 
 var PouchDB = process.browser ? global.PouchDB : require('pouchdb')
+var hoodieApi = require('../../')
 var uuid = PouchDB.utils.uuid
 
 module.exports = function (name) {
@@ -11,6 +12,8 @@ module.exports = function (name) {
   } : {
     db: require('memdown')
   }
+
+  PouchDB.plugin(hoodieApi)
 
   return new PouchDB(name, options)
 }
